aws_elasticache: Add AWS elasticache memcached client
This commit add the elasticache memcached client.
The setup is a bit complex, it need to be compiled and AWS doesn't really maintain this library.
It's working well on debian (tested with molecule) but doesn't work with ubuntu.

A parameter in the default can be overrided to setup this library easily.